Adding DeleteBtn Functionality

+++ routes/ideas.js
@@ -53,17 +53,25 @@ router.post('/', async (req, res) => {
 // To Update an idea we can use put request-
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(     //takes three arguments
-            req.params.id,
-            {
-                $set : {
-                    text : req.body.text,
-                    tag : req.body.tag,
-                }
-            },
-            {new : true}
-        );
-        res.json({success : true, data : updatedIdea});
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the username-
+        if(idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(     //takes three arguments
+                req.params.id,
+                {
+                    $set : {
+                        text : req.body.text,
+                        tag : req.body.tag,
+                    }
+                },
+                {new : true}
+            );
+            return res.json({success : true, data : updatedIdea});
+        }
+
+        // if Username do not match-
+        res.status(403).json({success : false, error : 'You are not authorized to update this resource.'});
     }
     catch(error) {
         console.log(error);
@@ -74,8 +82,16 @@ router.put('/:id', async (req, res) => {
 // Now to delete an idea using DELETE request-
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({success : true, data : {} });
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the username-
+        if(idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({success : true, data : {} });
+        }
+
+        // If username do not match--
+        res.status(403).json({success : false, error : 'You are not authorized to delete this resource.'});
     }
     catch(error) {
         console.log(error);
